backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "drf-nested-routers>=0.95.0",
    "pillow>=11.0.0",
    # Performance
    "django-redis>=5.4.0",
    "redis>=5.0.0",
    "orjson>=3.10.0",
    # Production
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.2.0",
    "whitenoise>=6.8.0",
]

[project.optional-dependencies]
dev = [
    "django-debug-toolbar>=4.4.0",
    "django-silk>=5.3.0",
]
